refactor: split monolithic xaml-converter into focused modules for better maintainability (code fails)
- Extract XML parsing logic into xaml-element-parser.ts
- Move formatting utilities to xaml-formatting-utils.ts
- Create xaml-element-processors.ts for core processing logic
- Maintain original public API in refactored xaml-converter.ts
